Move context processors from mozorg to base - no tests to move over
